snippets: Restructure internally to enable template docs
